+++ src/main/java/com/authlete/common/dto/AuthzDetailsElementSerializer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Authlete, Inc.
+ * Copyright (C) 2019-2021 Authlete, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -67,6 +67,7 @@ implements JsonSerializer<AuthzDetailsElement>
         jobject.add(        "actions",    fromStringArray(element.getActions()));
         jobject.add(        "datatypes",  fromStringArray(element.getDataTypes()));
         jobject.addProperty("identifier", element.getIdentifier());
+        jobject.add(        "privileges", fromStringArray(element.getPrivileges()));
 
         return jobject;
     }